requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiohttp-retry==2.9.1
aiosignal==1.4.0
amqp==5.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.10.0
appdirs==1.4.4
asttokens==3.0.0
async-timeout==5.0.1
asyncssh==2.21.0
atpublic==6.0.1
attrs==25.3.0
billiard==4.2.1
celery==5.5.3
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
comm==0.2.3
configobj==5.0.9
contourpy==1.3.2
cryptography==45.0.6
cycler==0.12.1
debugpy==1.8.16
decorator==5.2.1
dictdiffer==0.9.0
diskcache==5.6.3
distro==1.9.0
dpath==2.2.0
dulwich==0.24.1
dvc==3.62.0
dvc-data==3.16.12
dvc-http==2.32.0
dvc-objects==5.1.1
dvc-render==1.0.2
dvc-studio-client==0.22.0
dvc-task==0.40.2
entrypoints==0.4
exceptiongroup==1.3.0
executing==2.2.0
fastapi==0.116.1
filelock==3.19.1
flatten-dict==0.4.2
flufl-lock==8.2.0
fonttools==4.59.1
frozenlist==1.7.0
fsspec==2025.7.0
funcy==2.0
ghp-import==2.1.0
gitdb==4.0.12
gitpython==3.1.45
grandalf==0.8
gto==1.8.0
hydra-core==1.3.2
idna==3.10
iniconfig==2.1.0
ipykernel==6.30.1
ipython==8.37.0
iterative-telemetry==0.0.10
jedi==0.19.2
jinja2==3.1.6
joblib==1.5.1
jupyter-client==8.6.3
jupyter-core==5.8.1
kiwisolver==1.4.9
kombu==5.5.4
loguru==0.7.3
markdown==3.8.2
markdown-it-py==4.0.0
markupsafe==3.0.2
matplotlib==3.10.5
matplotlib-inline==0.1.7
mdurl==0.1.2
mergedeep==1.3.4
mkdocs==1.6.1
mkdocs-get-deps==0.2.0
multidict==6.6.4
mypy-extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.2.6
nvidia-nccl-cu12==2.27.7
omegaconf==2.3.0
orjson==3.11.2
packaging==25.0
pandas==2.3.1
pandera==0.26.0
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pillow==11.3.1
pip==25.2
platformdirs==4.3.8
pluggy==1.6.0
polars==1.32.3
prompt-toolkit==3.0.51
propcache==0.3.2
psutil==7.0.0
ptyprocess==0.7.0
pure-eval==0.2.3
pycparser==2.22
pydantic==2.11.7
pydantic-core==2.33.2
pydot==4.0.1
pygit2==1.18.2
pygments==2.19.2
pygtrie==2.5.0
pyparsing==3.2.3
pytest==8.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
pyyaml==6.0.2
pyyaml-env-tag==1.1
pyzmq==27.0.1
requests==2.32.5
rich==14.1.0
ruamel-yaml==0.18.15
ruff==0.12.9
scikit-learn==1.7.1
scipy==1.15.3
scmrepo==3.5.2
seaborn==0.13.2
semver==3.0.4
setuptools==80.9.0
shellingham==1.5.4
shortuuid==1.0.13
shtab==1.7.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sqltrie==0.11.2
stack-data==0.6.3
starlette==0.47.2
tabulate==0.9.0
threadpoolctl==3.6.0
-e file:///home/shahroz/ticket_urgency_classifier
tomli==2.2.1
tomlkit==0.13.3
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
typeguard==4.4.4
typer==0.16.1
typing-extensions==4.14.1
typing-inspect==0.9.0
typing-inspection==0.4.1
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
voluptuous==0.15.2
watchdog==6.0.0
wcwidth==0.2.13
xgboost==3.0.4
lightgbm==4.5.0
catboost==1.3.3
imbalanced-learn==0.13.1
yarl==1.20.1
zc-lockfile==3.0.post1
category_encoders

mlflow
